# frontend/openapi.json
{"openapi":"3.0.2","info":{"title":"lms","description":"lms API","version":"0.1.0"},"paths":{"/api/v1/users":{"get":{"tags":["users"],"summary":"Get Users","operationId":"get_users","parameters":[{"required":false,"schema":{"title":"Skip","type":"integer","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Users Api V1 Users Get","type":"array","items":{"$ref":"#/components/schemas/UserRead"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/books":{"get":{"tags":["books"],"summary":"Get Books","operationId":"get_books","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Books Api V1 Books Get","type":"array","items":{"$ref":"#/components/schemas/Book"}}}}}}},"post":{"tags":["books"],"summary":"Create Book","operationId":"create_book","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/books/{book_id}":{"get":{"tags":["books"],"summary":"Get Book","operationId":"get_book","parameters":[{"required":true,"schema":{"title":"Book Id","type":"integer"},"name":"book_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["books"],"summary":"Update Book","operationId":"update_book","parameters":[{"required":true,"schema":{"title":"Book Id","type":"integer"},"name":"book_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["books"],"summary":"Delete Book","operationId":"delete_book","parameters":[{"required":true,"schema":{"title":"Book Id","type":"integer"},"name":"book_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Delete Book Api V1 Books  Book Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/libraries":{"get":{"tags":["libraries"],"summary":"Get Libraries","operationId":"get_libraries","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Libraries Api V1 Libraries Get","type":"array","items":{"$ref":"#/components/schemas/Library"}}}}}}},"post":{"tags":["libraries"],"summary":"Create Library","operationId":"create_library","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LibraryCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Library"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/libraries/{library_id}":{"get":{"tags":["libraries"],"summary":"Get Library","operationId":"get_library","parameters":[{"required":true,"schema":{"title":"Library Id","type":"integer"},"name":"library_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Library"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["libraries"],"summary":"Update Library","operationId":"update_library","parameters":[{"required":true,"schema":{"title":"Library Id","type":"integer"},"name":"library_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LibraryUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Library"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["libraries"],"summary":"Delete Library","operationId":"delete_library","parameters":[{"required":true,"schema":{"title":"Library Id","type":"integer"},"name":"library_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Delete Library Api V1 Libraries  Library Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/transactions":{"get":{"tags":["transactions"],"summary":"Get Transactions","operationId":"get_transactions","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Transactions Api V1 Transactions Get","type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}}}},"post":{"tags":["transactions"],"summary":"Create Transaction","operationId":"create_transaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/transactions/{transaction_id}":{"get":{"tags":["transactions"],"summary":"Get Transaction","operationId":"get_transaction","parameters":[{"required":true,"schema":{"title":"Transaction Id","type":"integer"},"name":"transaction_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["transactions"],"summary":"Update Transaction","operationId":"update_transaction","parameters":[{"required":true,"schema":{"title":"Transaction Id","type":"integer"},"name":"transaction_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["transactions"],"summary":"Delete Transaction","operationId":"delete_transaction","parameters":[{"required":true,"schema":{"title":"Transaction Id","type":"integer"},"name":"transaction_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Delete Transaction Api V1 Transactions  Transaction Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/jwt/login":{"post":{"tags":["auth"],"summary":"Auth:Jwt.Login","operationId":"auth:jwt.login","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_auth_jwt_login_api_v1_auth_jwt_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BearerResponse"},"example":{"access_token":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoiOTIyMWZmYzktNjQwZi00MzcyLTg2ZDMtY2U2NDJjYmE1NjAzIiwiYXVkIjoiZmFzdGFwaS11c2VyczphdXRoIiwiZXhwIjoxNTcxNTA0MTkzfQ.M10bjOe45I5Ncu_uXvOmVV8QxnL-nZfcH96U90JaocI","token_type":"bearer"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"LOGIN_BAD_CREDENTIALS":{"summary":"Bad credentials or the user is inactive.","value":{"detail":"LOGIN_BAD_CREDENTIALS"}},"LOGIN_USER_NOT_VERIFIED":{"summary":"The user is not verified.","value":{"detail":"LOGIN_USER_NOT_VERIFIED"}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/jwt/logout":{"post":{"tags":["auth"],"summary":"Auth:Jwt.Logout","operationId":"auth:jwt.logout","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Missing token or inactive user."}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/auth/register":{"post":{"tags":["auth"],"summary":"Register:Register","operationId":"register:register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"REGISTER_USER_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"REGISTER_USER_ALREADY_EXISTS"}},"REGISTER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"REGISTER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/me":{"get":{"tags":["users"],"summary":"Users:Current User","operationId":"users:current_user","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"tags":["users"],"summary":"Users:Patch Current User","operationId":"users:patch_current_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"UPDATE_USER_EMAIL_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"UPDATE_USER_EMAIL_ALREADY_EXISTS"}},"UPDATE_USER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"UPDATE_USER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/users/{id}":{"get":{"tags":["users"],"summary":"Users:User","operationId":"users:user","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["users"],"summary":"Users:Delete User","operationId":"users:delete_user","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"responses":{"204":{"description":"Successful Response"},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"tags":["users"],"summary":"Users:Patch User","operationId":"users:patch_user","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"UPDATE_USER_EMAIL_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"UPDATE_USER_EMAIL_ALREADY_EXISTS"}},"UPDATE_USER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"UPDATE_USER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"BearerResponse":{"title":"BearerResponse","required":["access_token","token_type"],"type":"object","properties":{"access_token":{"title":"Access Token","type":"string"},"token_type":{"title":"Token Type","type":"string"}}},"Body_auth_jwt_login_api_v1_auth_jwt_login_post":{"title":"Body_auth_jwt_login_api_v1_auth_jwt_login_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"Book":{"title":"Book","required":["title","author","isbn","publication_date","genre","id"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"author":{"title":"Author","type":"string"},"isbn":{"title":"Isbn","type":"string"},"publication_date":{"title":"Publication Date","type":"string"},"genre":{"title":"Genre","type":"string"},"id":{"title":"Id","type":"integer"}}},"BookCreate":{"title":"BookCreate","required":["title","author","isbn","publication_date","genre"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"author":{"title":"Author","type":"string"},"isbn":{"title":"Isbn","type":"string"},"publication_date":{"title":"Publication Date","type":"string"},"genre":{"title":"Genre","type":"string"}}},"BookUpdate":{"title":"BookUpdate","required":["title","author","isbn","publication_date","genre"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"author":{"title":"Author","type":"string"},"isbn":{"title":"Isbn","type":"string"},"publication_date":{"title":"Publication Date","type":"string"},"genre":{"title":"Genre","type":"string"}}},"ErrorModel":{"title":"ErrorModel","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","anyOf":[{"type":"string"},{"type":"object","additionalProperties":{"type":"string"}}]}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Library":{"title":"Library","required":["name","location","id"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"location":{"title":"Location","type":"string"},"id":{"title":"Id","type":"integer"}}},"LibraryCreate":{"title":"LibraryCreate","required":["name","location"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"location":{"title":"Location","type":"string"}}},"LibraryUpdate":{"title":"LibraryUpdate","required":["name","location"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"location":{"title":"Location","type":"string"}}},"Transaction":{"title":"Transaction","required":["book_id","user_id","library_id","id","checkout_date","due_date"],"type":"object","properties":{"book_id":{"title":"Book Id","type":"integer"},"user_id":{"title":"User Id","type":"string","format":"uuid4"},"library_id":{"title":"Library Id","type":"integer"},"id":{"title":"Id","type":"integer"},"checkout_date":{"title":"Checkout Date","type":"string","format":"date"},"due_date":{"title":"Due Date","type":"string","format":"date"}}},"TransactionCreate":{"title":"TransactionCreate","required":["book_id","user_id","library_id"],"type":"object","properties":{"book_id":{"title":"Book Id","type":"integer"},"user_id":{"title":"User Id","type":"string","format":"uuid4"},"library_id":{"title":"Library Id","type":"integer"}}},"TransactionUpdate":{"title":"TransactionUpdate","required":["book_id","user_id","library_id","status","return_date"],"type":"object","properties":{"book_id":{"title":"Book Id","type":"integer"},"user_id":{"title":"User Id","type":"string","format":"uuid4"},"library_id":{"title":"Library Id","type":"integer"},"status":{"title":"Status","type":"string"},"return_date":{"title":"Return Date","type":"string","format":"date"}}},"UserCreate":{"title":"UserCreate","required":["email","password"],"type":"object","properties":{"email":{"title":"Email","type":"string","format":"email"},"password":{"title":"Password","type":"string"},"is_active":{"title":"Is Active","type":"boolean","default":true},"is_superuser":{"title":"Is Superuser","type":"boolean","default":false},"is_verified":{"title":"Is Verified","type":"boolean","default":false}}},"UserRead":{"title":"UserRead","required":["email"],"type":"object","properties":{"id":{"title":"Id"},"email":{"title":"Email","type":"string","format":"email"},"is_active":{"title":"Is Active","type":"boolean","default":true},"is_superuser":{"title":"Is Superuser","type":"boolean","default":false},"is_verified":{"title":"Is Verified","type":"boolean","default":false}},"description":"Base User model."},"UserUpdate":{"title":"UserUpdate","type":"object","properties":{"password":{"title":"Password","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"is_active":{"title":"Is Active","type":"boolean"},"is_superuser":{"title":"Is Superuser","type":"boolean"},"is_verified":{"title":"Is Verified","type":"boolean"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/jwt/login"}}}}}}